trash menu order

+++ fos-client/src/app/components/order-detail/list-ordered-foods/list-ordered-foods.component.ts
@@ -1,7 +1,19 @@
-import { Component, OnInit, ViewChild, Input } from "@angular/core";
+import {
+  Component,
+  OnInit,
+  ViewChild,
+  Input,
+  Output,
+  EventEmitter
+} from "@angular/core";
 import { User } from "src/app/models/user";
 import { Order } from "src/app/models/order";
-import { MatSort, MatPaginator, MatTableDataSource } from "@angular/material";
+import {
+  MatSort,
+  MatPaginator,
+  MatTableDataSource,
+  MatDialog
+} from "@angular/material";
 import { Food } from "src/app/models/food";
 import { FoodDetailJson } from "src/app/models/food-detail-json";
 import {
@@ -10,6 +22,8 @@ import {
 } from "html2canvas/dist/types/css/property-descriptors/float";
 import { Event } from "src/app/models/event";
 import moment from "moment";
+import { DialogCheckActionComponent } from "./dialog-check-action/dialog-check-action.component";
+import { Overlay } from "@angular/cdk/overlay";
 
 @Component({
   selector: "app-list-ordered-foods",
@@ -23,10 +37,18 @@ export class ListOrderedFoodsComponent implements OnInit {
   @Input() event: Event;
   @Input() order: Order;
   @Input("isOrder") isOrder: boolean;
+  @Output() valueChange = new EventEmitter<FoodDetailJson>();
 
   @ViewChild(MatSort, { static: true }) sort: MatSort;
   foodOrdered: Food;
-  displayedColumns2: string[] = ["name", "price", "amount", "total", "comment"];
+  displayedColumns2: string[] = [
+    "name",
+    "price",
+    "amount",
+    "total",
+    "comment",
+    "trash"
+  ];
   dataSource2: MatTableDataSource<FoodDetailJson>;
   public FoodOfAmount: any = {};
   day: number;
@@ -34,7 +56,7 @@ export class ListOrderedFoodsComponent implements OnInit {
   year: number;
   hour: number;
   minutes: number;
-  constructor() {}
+  constructor(public dialog: MatDialog, private overlay: Overlay) {}
   load = true;
   @Input() totalBudget: Number;
   ngOnInit() {
@@ -64,7 +86,9 @@ export class ListOrderedFoodsComponent implements OnInit {
   }
   DeleteFoodDetail(food: Food) {
     var deleteItem = this.dataSource2.data.findIndex(x => x.IdFood == food.Id);
-    this.dataSource2.data.splice(deleteItem, 1);
+    if (deleteItem >= 0) {
+      this.dataSource2.data.splice(deleteItem, 1);
+    }
     this.dataSource2.filter = "";
   }
   numberWithCommas(x: Number) {
@@ -116,5 +140,24 @@ export class ListOrderedFoodsComponent implements OnInit {
     this.dataSource2.data = this.order.FoodDetail;
     this.dataSource2.sort = this.sort;
     this.load = false;
+  }
+  openDialog(row: FoodDetailJson): void {
+    const dialogRef = this.dialog.open(DialogCheckActionComponent, {
+      scrollStrategy: this.overlay.scrollStrategies.noop(),
+      autoFocus: false,
+      maxHeight: "98vh",
+      width: "80%",
+      data: row
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result.Id != undefined) {
+        var deleteItem = this.dataSource2.data.findIndex(
+          x => x.IdFood == result.Id
+        );
+        this.dataSource2.data.splice(deleteItem, 1);
+        this.dataSource2.filter = "";
+        this.valueChange.emit(row);
+      }
+    });
   }
 }
